refactor: dom optimization
-Move responsive optimization code into it's own hook.

Add hook that checks window width to determine how many elements should be displayed on screen.
